# src/main/resources/swagger/swagger-workstation.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: API Workstation
  description: 'Hackathon'
  contact:
    name: Team 3 - GFT Brasil
    email: andu@gft.com
host: gft.com.br
basePath: /estacoes/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /booking:
    post:
      tags:
        - Bookings
      summary: Create a booking
      description: 'put a user'
      operationId: createBooking
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: Booking
          in: body
          description: Booking object
          required: true
          schema:
            $ref: '#/definitions/Booking'
      responses:
        '201':
          description: Create new resource
          schema:
            $ref: '#/definitions/BookingResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /booking/{idUser}/user:
    get:
      tags:
        - Bookings
      summary: Gets a booking
      description: 'get a booking'
      operationId: getBookingsByCodeUser
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: idUser
          in: path
          description: Code of booking
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/BookingResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{id}/checkin:
    patch:
      tags:
        - Bookings
      summary: Update a user
      description: 'put a user'
      operationId: createCheckin
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: booking
          required: true
          type: string
          x-example: 'XPTO'
        - name: Booking
          in: body
          description: data of the Booking
          required: true
          schema:
            $ref: '#/definitions/Booking'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{id}/checkout:
    patch:
      tags:
        - Bookings
      summary: Update a user
      description: 'put a user'
      operationId: createCheckout
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of the Booking
          required: true
          type: string
          x-example: 'XPTO'
        - name: Booking
          in: body
          description: data of the Booking
          required: true
          schema:
            $ref: '#/definitions/Booking'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
...
securityDefinitions:
  OauthSecurity:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    tokenUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    scopes:
      bearer: Client authorization token - Bearer JWT
security:
  - OauthSecurity:
      - bearer
definitions:
...
  Booking:
    description: Booking
    type: object
    properties:
      codeUser:
        type: string
        example: string
        description: Name of Booking
      codeWorkstation:
        type: string
        example: string
        description: Name of Booking
      endereco:
        type: string
        example: andu
        description: Unique identifier of the Booking
      bookingDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Start Date of the validity of the Booking.
      checkinDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkin.
      checkoutDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkou.
      sizeHours:
        type: string
        format: int
        example: yyyy-mm-dd
        description: Start Date of the validity of the Booking.
      status:
        type: string
        enum: [ACTIVE, INACTIVE, CANCEL]
  BookingResponse:
    type: object
    properties:
      bookings:
        type: array
        items:
          $ref: '#/definitions/Booking'
...
  ErrorModel:
    type: object
    required:
      - codeBuilderComplete
      - message
    properties:
      code:
        type: string
      message:
        type: string